# docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: BigSchool API
  version: 0.2.0
  description: |
    API REST para la tienda online de zapatillas BigSchool.

    ## Stack
    - **Backend**: NestJS (Node.js + TypeScript)
    - **Base de datos**: Supabase (PostgreSQL gestionado)
    - **Imágenes**: Cloudinary (CDN + transformaciones on-the-fly)
    - **Pagos**: Stripe

    ## Autenticación
    Gestionada por **Supabase Auth** en el cliente Angular (email/password y Google OAuth).
    El cliente obtiene un JWT de Supabase y lo envía como `Authorization: Bearer <token>`.
    NestJS valida el JWT contra la clave pública de Supabase — no genera tokens propios.

    ## Roles
    - `USER` — usuario registrado (asignado por defecto)
    - `ADMIN` — acceso completo incluyendo rutas `/admin/*`

    ## Imágenes (Cloudinary)
    Todas las URLs de imagen son URLs de Cloudinary.
    Transformaciones on-the-fly: añadir parámetros en la URL → `f_auto,q_auto,w_400`.

    ## Flujo de pago (Stripe)
    1. `POST /orders` → crea pedido en estado `pending_payment`
    2. `POST /payments/intent` → crea PaymentIntent, devuelve `clientSecret`
    3. Frontend completa pago con Stripe.js
    4. `POST /webhooks/stripe` → actualiza estado del pedido a `confirmed`
  contact:
    email: javiersalvadormarco@gmail.com
servers:
  - url: http://localhost:3000/api/v1
    description: Desarrollo local
  - url: https://api.bigschool.com/v1
    description: Producción
...
paths:
...
  /products:
    get:
      tags: [products]
      summary: Listar productos con filtros y paginación
      operationId: listProducts
      parameters:
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - name: categoryId
          in: query
          schema: { type: string, format: uuid }
        - name: brandId
          in: query
          schema: { type: string, format: uuid }
        - name: gender
          in: query
          schema:
            $ref: '#/components/schemas/Gender'
        - name: minPrice
          in: query
          schema: { type: number, minimum: 0 }
        - name: maxPrice
          in: query
          schema: { type: number }
        - name: size
          in: query
          schema: { type: number }
          description: Filtrar por talla EU disponible en stock
        - name: color
          in: query
          schema: { type: string }
        - name: search
          in: query
          schema: { type: string }
          description: Búsqueda por nombre o descripción
        - name: onSale
          in: query
          schema: { type: boolean }
          description: Solo productos en oferta
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [price_asc, price_desc, newest, popularity, rating]
            default: newest
      responses:
        '200':
          description: Productos paginados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductPage'
  /products/{productId}:
    parameters:
      - $ref: '#/components/parameters/ProductId'
    get:
      tags: [products]
      summary: Detalle de producto (con variantes y SEO)
      operationId: getProductById
      responses:
        '200':
          description: Producto detallado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDetail'
        '404':
          $ref: '#/components/responses/NotFound'
  /products/{productId}/related:
    parameters:
      - $ref: '#/components/parameters/ProductId'
    get:
      tags: [products]
      summary: Productos relacionados
      operationId: getRelatedProducts
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 20, default: 8 }
      responses:
        '200':
          description: Productos relacionados
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        '404':
          $ref: '#/components/responses/NotFound'
  /products/{productId}/reviews:
    parameters:
      - $ref: '#/components/parameters/ProductId'
    get:
      tags: [reviews]
      summary: Listar valoraciones del producto
      operationId: listProductReviews
      parameters:
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 50, default: 10 }
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [newest, highest_rating, lowest_rating, most_helpful]
            default: newest
      responses:
        '200':
          description: Valoraciones paginadas con resumen estadístico
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewPage'
    post:
      tags: [reviews]
      summary: Publicar valoración
      description: El usuario debe haber comprado el producto (compra verificada).
      operationId: createReview
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReviewRequest'
      responses:
        '201':
          description: Valoración publicada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /products/{productId}/variants/{variantId}/notify-stock:
    parameters:
      - $ref: '#/components/parameters/ProductId'
      - $ref: '#/components/parameters/VariantId'
    post:
      tags: [products]
      summary: Suscribirse a notificación de disponibilidad
      operationId: subscribeStockNotification
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Suscripción creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockNotification'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    delete:
      tags: [products]
      summary: Cancelar suscripción de disponibilidad
      operationId: unsubscribeStockNotification
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Suscripción cancelada
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ===========================================================================
  # REVIEWS (gestión propia)
  # ===========================================================================
  /reviews/{reviewId}:
    parameters:
      - $ref: '#/components/parameters/ReviewId'
    patch:
      tags: [reviews]
      summary: Editar mi valoración
      operationId: updateReview
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReviewRequest'
      responses:
        '200':
          description: Valoración actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [reviews]
      summary: Eliminar mi valoración
      operationId: deleteReview
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Valoración eliminada
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ===========================================================================
  # WISHLIST
  # ===========================================================================
...
  /admin/products/{productId}:
    parameters:
      - $ref: '#/components/parameters/ProductId'
    get:
      tags: [admin]
      summary: '[Admin] Detalle de producto'
      operationId: adminGetProduct
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Producto con métricas de ventas y stock
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminProductDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [admin]
      summary: '[Admin] Actualizar producto'
      operationId: adminUpdateProduct
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProductRequest'
      responses:
        '200':
          description: Producto actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminProductDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
    delete:
      tags: [admin]
      summary: '[Admin] Desactivar producto (soft delete)'
      operationId: adminDeleteProduct
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Producto desactivado
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT emitido por Supabase Auth
  parameters:
    ProductId:
      name: productId
      in: path
      required: true
      schema: { type: string, format: uuid }
    VariantId:
      name: variantId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    ReviewId:
      name: reviewId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    BadRequest:
      description: Petición incorrecta
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: No autenticado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Sin permisos suficientes
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Conflicto de recursos
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: Error de validación
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
  schemas:
...
    Gender:
      type: string
      enum: [man, woman, unisex, kids]
...
    Category:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        imageUrl:
          type: string
          format: uri
          nullable: true
        productCount:
          type: integer
...
    Brand:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        logoUrl:
          type: string
          format: uri
          nullable: true
...
    ProductVariant:
      type: object
      properties:
        id:
          type: string
          format: uuid
        sku:
          type: string
        size:
          type: number
          description: Talla EU
        color:
          type: string
        colorHex:
          type: string
          example: '#1a1a1a'
        stock:
          type: integer
          minimum: 0
        imageUrls:
          type: array
          items:
            type: string
            format: uri
          description: URLs de Cloudinary
    ProductSeo:
      type: object
      properties:
        metaTitle:
          type: string
          maxLength: 60
        metaDescription:
          type: string
          maxLength: 160
        canonicalUrl:
          type: string
          format: uri
    Product:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        brand:
          $ref: '#/components/schemas/Brand'
        category:
          $ref: '#/components/schemas/Category'
        gender:
          $ref: '#/components/schemas/Gender'
        price:
          type: number
          format: float
          description: Precio base en EUR
        salePrice:
          type: number
          format: float
          nullable: true
          description: Precio de oferta en EUR
        saleEndsAt:
          type: string
          format: date-time
          nullable: true
        thumbnailUrl:
          type: string
          format: uri
        averageRating:
          type: number
          format: float
          minimum: 0
          maximum: 5
          nullable: true
        reviewCount:
          type: integer
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    ProductDetail:
      allOf:
        - $ref: '#/components/schemas/Product'
        - type: object
          properties:
            description:
              type: string
            variants:
              type: array
              items:
                $ref: '#/components/schemas/ProductVariant'
            seo:
              $ref: '#/components/schemas/ProductSeo'
    AdminProductDetail:
      allOf:
        - $ref: '#/components/schemas/ProductDetail'
        - type: object
          properties:
            totalStock:
              type: integer
            totalSold:
              type: integer
    ProductPage:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Product'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    UpdateProductRequest:
      type: object
      properties:
        name:
          type: string
        brandId:
          type: string
          format: uuid
        categoryId:
          type: string
          format: uuid
        gender:
          $ref: '#/components/schemas/Gender'
        price:
          type: number
          format: float
          minimum: 0
        salePrice:
          type: number
          format: float
          nullable: true
        saleEndsAt:
          type: string
          format: date-time
          nullable: true
        description:
          type: string
        thumbnailUrl:
          type: string
          format: uri
        isActive:
          type: boolean
        seo:
          $ref: '#/components/schemas/ProductSeo'
...
    Review:
      type: object
      properties:
        id:
          type: string
          format: uuid
        productId:
          type: string
          format: uuid
        user:
          type: object
          properties:
            id:
              type: string
              format: uuid
            firstName:
              type: string
            avatarUrl:
              type: string
              format: uri
              nullable: true
        rating:
          type: integer
          minimum: 1
          maximum: 5
        title:
          type: string
          nullable: true
        body:
          type: string
        verifiedPurchase:
          type: boolean
        helpfulCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    CreateReviewRequest:
      type: object
      required: [rating, body]
      properties:
        rating:
          type: integer
          minimum: 1
          maximum: 5
        title:
          type: string
          maxLength: 100
          nullable: true
        body:
          type: string
          minLength: 10
          maxLength: 2000
    ReviewPage:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Review'
        summary:
          type: object
          properties:
            averageRating:
              type: number
              format: float
            totalCount:
              type: integer
            distribution:
              type: object
              description: 'Conteo por estrella: { "1": 5, "2": 3, "3": 10, "4": 22, "5": 48 }'
              additionalProperties:
                type: integer
        meta:
          $ref: '#/components/schemas/PaginationMeta'

    # ─── STOCK NOTIFICATIONS ─────────────────────────────────────────────────
    StockNotification:
      type: object
      properties:
        id:
          type: string
          format: uuid
        productId:
          type: string
          format: uuid
        variantId:
          type: string
          format: uuid
        productName:
          type: string
        size:
          type: number
        color:
          type: string
        createdAt:
          type: string
          format: date-time

    # ─── SIZE GUIDE ──────────────────────────────────────────────────────────
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
        message:
          type: string
        error:
          type: string
    ValidationErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 422
        message:
          type: array
          items:
            type: string
        error:
          type: string
          example: Unprocessable Entity